+++ netket/driver/vmc.py
@@ -28,7 +28,7 @@ from netket.optimizer import (
 from netket.utils import warn_deprecation
 
 
-from .vmc_common import info, apply_preconditioner
+from .vmc_common import info, ensure_step_value
 from .abstract_variational_driver import AbstractVariationalDriver
 
 
@@ -112,7 +112,7 @@ class VMC(AbstractVariationalDriver):
 
         self._ham = hamiltonian.collect()  # type: AbstractOperator
 
-        self.preconditioner = preconditioner
+        self.preconditioner = ensure_step_value(preconditioner)
 
         self._dp = None  # type: PyTree
         self._S = None
@@ -133,7 +133,9 @@ class VMC(AbstractVariationalDriver):
 
         # if it's the identity it does
         # self._dp = self._loss_grad
-        apply_preconditioner(self)
+        self._dp = self.preconditioner(
+            self.state, self._loss_grad, step_value=self.step_count
+        )
 
         # If parameters are real, then take only real part of the gradient (if it's complex)
         self._dp = jax.tree_multimap(